+++ source/patches.hpp
@@ -6,10 +6,10 @@
 class Patches
 {
   auto static constexpr refresh_rate_ms_ = 20;
-  auto static constexpr game_running_address_ = uintptr_t{ 0x00400000 };
-  auto static constexpr player_state_offset_ = uintptr_t{ 0x530 };
-  auto static constexpr health_offset_ = uintptr_t{ 0x540 };
-  auto static constexpr armour_offset_ = uintptr_t{ 0x548 };
+  auto static constexpr game_running_address_ = uint32_t{ 0x00400000 };
+  auto static constexpr player_state_offset_ = uint32_t{ 0x530 };
+  auto static constexpr health_offset_ = uint32_t{ 0x540 };
+  auto static constexpr armour_offset_ = uint32_t{ 0x548 };
   auto static constexpr max_health_target_ = 5.682f;
   auto static constexpr max_health_target_actual_ = 1.0001f;
   auto static constexpr max_armour_target_ = 0.0f;
@@ -18,8 +18,8 @@ class Patches
 
 public:
   auto static IsGameRunning() -> bool;
-  auto static PatchHealthBar(int32_t ver_offset) -> bool;
-  auto static StartOHKOThread(int32_t ver_offset, bool die_after_bike_fall)
+  auto static PatchHealthBar(uint32_t ver_offset) -> bool;
+  auto static StartOHKOThread(uint32_t ver_offset, bool die_after_bike_fall)
     -> bool;
 
 private:
